@@ -7,11 +7,11 @@ module ActionDispatch::Routing
 
     def devise_otp(mapping, controllers)
 
-      namespace :otp, :module => :devise_otp do
+      namespace :otp, module: 'devise/otp' do
         resource :token, :only => [:show, :update, :destroy],
-                 :path => mapping.path_names[:token], :controller => controllers[:otp_tokens] do
+          :path => mapping.path_names[:token], :controller => controllers[:otp_tokens] do
 
-          if Devise.otp_trust_persistence
+          if ::Devise.otp_trust_persistence
             get  :persistence, :action => 'get_persistence'
             post :persistence, :action => 'clear_persistence'
             delete :persistence, :action => 'delete_persistence'
@@ -21,7 +21,7 @@ module ActionDispatch::Routing
         end
 
         resource :credential, :only => [:show, :update],
-                 :path => mapping.path_names[:credentials], :controller => controllers[:otp_credentials] do
+          :path => mapping.path_names[:credentials], :controller => controllers[:otp_credentials] do
 
           get  :refresh, :action => 'get_refresh'
           put :refresh, :action => 'set_refresh'
