@@ -1,5 +1,5 @@
 module Devise
   module Otp
-    VERSION = "0.3.5"
+    VERSION = "0.3.6"
   end
 end
