@@ -6,6 +6,7 @@ class DeviseOtpAddTo<%= table_name.camelize %> < ActiveRecord::Migration
       t.boolean   :otp_enabled,          :default => false, :null => false
       t.boolean   :otp_mandatory,        :default => false, :null => false
       t.datetime  :otp_enabled_on
+      t.integer   :otp_time_drift,       :default => 0, :null => false
       t.integer   :otp_failed_attempts,  :default => 0, :null => false
       t.integer   :otp_recovery_counter, :default => 0, :null => false
       t.string    :otp_persistence_seed
@@ -16,7 +17,7 @@ class DeviseOtpAddTo<%= table_name.camelize %> < ActiveRecord::Migration
     add_index :<%= table_name %>, :otp_session_challenge,  :unique => true
     add_index :<%= table_name %>, :otp_challenge_expires
   end
-  
+
   def self.down
     change_table :<%= table_name %> do |t|
       t.remove :otp_auth_secret, :otp_recovery_secret, :otp_enabled, :otp_mandatory, :otp_enabled_on, :otp_session_challenge,
