@@ -1,4 +1,4 @@
 ActiveRecord::Migration.verbose = false
 ActiveRecord::Base.logger = Logger.new(nil)
 
-ActiveRecord::Migrator.migrate(File.expand_path("../../dummy/db/migrate/", __FILE__))
+ActiveRecord::MigrationContext.new(File.expand_path("../../dummy/db/migrate/", __FILE__)).migrate
